EarthSciMLBase

Documentation for EarthSciMLBase.

This package contains utilities for constructing Earth Science models in Julia using ModelingToolkit.jl.

EarthSciMLBase.AbstractEarthSciMLSystem

EarthSciMLBase.Advection

EarthSciMLBase.ComposedEarthSciMLSystem

EarthSciMLBase.ConnectorSystem

EarthSciMLBase.ConstantWind

EarthSciMLBase.EarthSciMLODESystem

EarthSciMLBase.ICBC

EarthSciMLBase.ICBCcomponent

EarthSciMLBase.MeanWind

EarthSciMLBase.constBC

EarthSciMLBase.constIC

EarthSciMLBase.periodicBC

EarthSciMLBase.zerogradBC

EarthSciMLBase.add_dims

EarthSciMLBase.dims

EarthSciMLBase.domains

EarthSciMLBase.get_mtk

EarthSciMLBase.ivar

EarthSciMLBase.operator_compose

EarthSciMLBase.pvars

EarthSciMLBase.AbstractEarthSciMLSystem — Type

One or more ModelingToolkit systems of equations. EarthSciML uses custom types to allow automatic composition of different systems together.

EarthSciMLBase.Advection — Type

Apply advection to a model.

Example

using EarthSciMLBase
using DomainSets, MethodOfLines, ModelingToolkit, Plots

# Create our independent variable `t` and our partially-independent variable `x`.
@parameters t, x

# Create our ODE system of equations as a subtype of `EarthSciMLODESystem`.
# Creating our system in this way allows us to convert it to a PDE system 
# using just the `+` operator as shown below.
struct ExampleSys <: EarthSciMLODESystem
    sys::ODESystem
    function ExampleSys(t; name)
        @variables y(t)
        @parameters p=2.0
        D = Differential(t)
        new(ODESystem([D(y) ~ p], t; name))
    end
end
@named sys = ExampleSys(t)

# Create our initial and boundary conditions.
icbc = ICBC(constBC(1.0, x ∈ Interval(0, 1.0)), constIC(0.0, t ∈ Interval(0, 1.0)))

# Convert our ODE system to a PDE system and add advection to each of the state variables.
# We're also adding a constant wind in the x-direction, with a speed of 1.0.
sys_advection = sys + icbc + ConstantWind(t, 1.0) + Advection()
sys_mtk = get_mtk(sys_advection)

# Discretize the system and solve it.
discretization = MOLFiniteDifference([x=>10], t, approx_order=2)
@time prob = discretize(sys_mtk, discretization)
@time sol = solve(prob, Tsit5(), saveat=0.1)

# Plot the solution.
discrete_x = sol[x]
discrete_t = sol[t]
@variables sys₊y(..)
soly = sol[sys₊y(x, t)]
anim = @animate for k in 1:length(discrete_t)
    plot(soly[1:end, k], title="t=$(discrete_t[k])", ylim=(0,2.5), lab=:none)
end
gif(anim, fps = 8)

EarthSciMLBase.ComposedEarthSciMLSystem — Type

A system for composing together other systems using the + operator.

The easiest way to explain how this works and why we might want to do it is with an example. The following example is modified from the ModelingToolkit.jl documentation here:

Example:

using EarthSciMLBase
using ModelingToolkit

# Set up our independent variable time, which will be shared by all systems.
@parameters t

# Create three systems which we will compose together.
struct SEqn <: EarthSciMLODESystem
    sys::ODESystem

    function SEqn(t) 
        @variables S(t), I(t), R(t)
        D = Differential(t)
        N = S + I + R
        @parameters β
        @named seqn = ODESystem([D(S) ~ -β*S*I/N])
        new(seqn)
    end
end

struct IEqn <: EarthSciMLODESystem
    sys::ODESystem

    function IEqn(t) 
        @variables S(t), I(t), R(t)
        D = Differential(t)
        N = S + I + R
        @parameters β,γ
        @named ieqn = ODESystem([D(I) ~ β*S*I/N-γ*I])
        new(ieqn)
    end
end

struct REqn <: EarthSciMLODESystem
    sys::ODESystem

    function REqn(t) 
        @variables I(t), R(t)
        D = Differential(t)
        @parameters γ
        @named reqn = ODESystem([D(R) ~ γ*I])
        new(reqn)
    end
end

# Create functions to allow us to compose the systems together using the `+` operator.
function Base.:(+)(s::SEqn, i::IEqn)::ComposedEarthSciMLSystem
    seqn = s.sys
    ieqn = i.sys
    ComposedEarthSciMLSystem(
        ConnectorSystem([
            ieqn.S ~ seqn.S,
            seqn.I ~ ieqn.I], s, i), 
        s, i,
    )
end

function Base.:(+)(s::SEqn, r::REqn)::ComposedEarthSciMLSystem
    seqn = s.sys
    reqn = r.sys
    ComposedEarthSciMLSystem(
        ConnectorSystem([seqn.R ~ reqn.R], s, r), 
        s, r,
    )
end

function Base.:(+)(i::IEqn, r::REqn)::ComposedEarthSciMLSystem
    ieqn = i.sys
    reqn = r.sys
    ComposedEarthSciMLSystem(
        ConnectorSystem([
            ieqn.R ~ reqn.R,
            reqn.I ~ ieqn.I], i, r), 
        i, r,
    )
end

# Instantiate our three systems.
seqn, ieqn, reqn = SEqn(t), IEqn(t), REqn(t)

# Compose the systems together using the `+` operator. This is the fancy part!
sir = seqn + ieqn + reqn

# Finalize the system for solving.
sirfinal = get_mtk(sir)

# Show the equations in our combined system.
equations(structural_simplify(sirfinal))

# output
3-element Vector{Equation}:
 Differential(t)(reqn₊R(t)) ~ reqn₊γ*ieqn₊I(t)
 Differential(t)(seqn₊S(t)) ~ (-seqn₊β*ieqn₊I(t)*seqn₊S(t)) / (ieqn₊I(t) + reqn₊R(t) + seqn₊S(t))
 Differential(t)(ieqn₊I(t)) ~ (ieqn₊β*ieqn₊I(t)*seqn₊S(t)) / (ieqn₊I(t) + reqn₊R(t) + seqn₊S(t)) - ieqn₊γ*ieqn₊I(t)

EarthSciMLBase.ConnectorSystem — Type

A connector for two systems.

eqs

from

to

EarthSciMLBase.ConstantWind — Type

Construct a constant wind velocity model component.

EarthSciMLBase.EarthSciMLODESystem — Type

A type for actual implementations of ODE systems.

EarthSciMLBase.ICBC — Type

ICBC(icbc::ICBCcomponent...)

Initial and boundary conditions for a ModelingToolkit.jl PDESystem. It can be used with the + operator to add initial and boundary conditions to a ModelingToolkit.jl ODESystem or Catalyst.jl ReactionSystem.

icbc: The sets of initial and/or boundary conditions.

Example:

using EarthSciMLBase
using ModelingToolkit, DomainSets

# Set up ODE system
@parameters x y t
@variables u(t) v(t)
Dt = Differential(t)

x_min = y_min = t_min = 0.0
x_max = y_max = 1.0
t_max = 11.5

eqs = [
    Dt(u) ~ √abs(v),
    Dt(v) ~ √abs(u),
]

@named sys = ODESystem(eqs)

# Create constant initial conditions = 16.0 and boundary conditions = 4.0.
icbc = ICBC(
    constBC(16.0, 
        x ∈ Interval(x_min, x_max),
        y ∈ Interval(y_min, y_max),
    ),
    constIC(4.0, t ∈ Interval(t_min, t_max)),
)

# Convert to PDESystem and add constant initial and boundary conditions.
pdesys = sys + icbc

pdesys.bcs

EarthSciMLBase.ICBCcomponent — Type

Initial and boundary condition components that can be combined to create an ICBC object.

EarthSciMLBase.MeanWind — Type

A model component that represents the mean wind velocity, where t is the independent variable and ndims is the number of dimensions that wind is traveling in.

EarthSciMLBase.constBC — Type

Construct constant boundary conditions equal to the value specified by val.

val: The value of the constant boundary conditions.

partialdomains: The partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

EarthSciMLBase.constIC — Type

Construct constant initial conditions equal to the value specified by val.

val: The value of the constant initial conditions.

indepdomain: The independent domain, e.g. t ∈ Interval(t_min, t_max).

EarthSciMLBase.periodicBC — Type

Construct periodic boundary conditions for the given partialdomains. Periodic boundary conditions are defined as when the value at one side of the domain is set equal to the value at the other side, so that the domain "wraps around" from one side to the other.

partialdomains: The partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

EarthSciMLBase.zerogradBC — Type

Construct zero-gradient boundary conditions for the given partialdomains.

partialdomains: The partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

EarthSciMLBase.add_dims — Method

add_dims(expression, vars, dims)
add_dims(equation, vars, dims)

Add the given dimensions to each variable in vars in the given expression or equation. Each variable in vars must be unidimensional, i.e. defined like @variables u(t) rather than @variables u(..).

Example:

using EarthSciMLBase, ModelingToolkit

@parameters x y k t
@variables u(t) q(t)
exp = 2u + 3k*q + 1
EarthSciMLBase.add_dims(exp, [u, q], [x, y, t])

# output
1 + 2u(x, y, t) + 3k*q(x, y, t)

EarthSciMLBase.dims — Method

dims(
    icbc::EarthSciMLBase.ICcomponent
) -> Vector{Symbolics.Num}

Returns the dimensions of the independent and partial domains associated with these initial or boundary conditions.

EarthSciMLBase.domains — Method

domains(icbc::EarthSciMLBase.ICcomponent) -> Vector

Returns the domains associated with these initial or boundary conditions.

EarthSciMLBase.get_mtk — Method

julia get_mtk( sys::AbstractEarthSciMLSystem ) -> ModelingToolkit.AbstractSystem

Return the ModelingToolkit version of this system.

EarthSciMLBase.ivar — Method

ivar(icbc::ICBC) -> Any

Return the independent variable associated with these initial and boundary conditions.

EarthSciMLBase.operator_compose — Method

operator_compose(
    a::EarthSciMLODESystem,
    b::EarthSciMLODESystem
) -> ComposedEarthSciMLSystem

Compose to systems of equations together by adding the right-hand side terms together of equations that have matching left-hand sides.

The example below shows that when we operator_compose two systems together that are both equal to D(x) = p, the resulting system is equal to D(x) = 2p.

Example

using EarthSciMLBase
using ModelingToolkit

@parameters t

struct ExampleSys <: EarthSciMLODESystem
    sys::ODESystem

    function ExampleSys(t; name)
        @variables x(t)
        @parameters p
        D = Differential(t)
        new(ODESystem([D(x) ~ p], t; name))
    end
end

@named sys1 = ExampleSys(t)
@named sys2 = ExampleSys(t)

combined = operator_compose(sys1, sys2)

combined_mtk = get_mtk(combined)
equations(structural_simplify(combined_mtk))

# The simplified equation should be D(x) = p + sys2_xˍt, where sys2_xˍt is also equal to p.
# output
1-element Vector{Equation}:
 Differential(t)(sys1₊x(t)) ~ sys1₊p + sys2₊sys2_xˍt(t)

EarthSciMLBase.pvars — Method

pvars(icbc::ICBC) -> Any

Return the partial independent variables associated with these initial and boundary conditions.