EarthSciMLBase

Documentation for EarthSciMLBase.

EarthSciMLBase.AddDims

Base.:+

Base.:+

EarthSciMLBase.add_dims

EarthSciMLBase.AddDims — Type

AddDims(dims...)

Construct an object that can add dimensions to the variables in a system of equations. The object can be used using the + operator.

Example:

using ModelingToolkit

@parameters x y t k
@variables u(t) q(t)
Dt = Differential(t)

eqs = [
    Dt(u) ~ 2u + 3k*q + 1, 
    Dt(q) ~ 3u + k*q + 1
]
@named sys = ODESystem(eqs)
sys + AddDims(x, y, t)

# output
2-element Vector{Equation}:
 Differential(t)(u(x, y, t)) ~ 1 + 2u(x, y, t) + 3k*q(x, y, t)
 Differential(t)(q(x, y, t)) ~ 1 + k*q(x, y, t) + 3u(x, y, t)

Base.:+ — Method

    a::Symbolics.Equation + b::Symbolics.Equation

Add two equations. If the left-hand-sides of the two equations are the same, a single equation is returned with the two right-hand-sides added together. If the left-hand-sides of the two equations are different, a vector of the two equations is returned.

Base.:+ — Method

    a::Vector{Symbolics.Equation} + b::Vector{Symbolics.Equation}

Add two vectors of equations using the same rules as when adding two individual equations.

EarthSciMLBase.add_dims — Method

add_dims(expression, vars, dims)
add_dims(equation, vars, dims)
add_dims(ModelingToolkit.AbstractSystem, dims)
add_dims(Catalyst.ReactionSystem, dims)

Add the given dimensions to each variable in vars in the given expression or system. Each variable in vars must be unidimensional, i.e. defined like @variables u(t) rather than @variables u(..).

Example:

using EarthSciMLBase, ModelingToolkit

@parameters x y k t
@variables u(t) q(t)
exp = 2u + 3k*q + 1
EarthSciMLBase.add_dims(exp, [u, q], x, y, t)

# output
1 + 2u(x, y, t) + 3k*q(x, y, t)