EarthSciMLBase

Documentation for EarthSciMLBase.

This package contains utilities for constructing Earth Science models in Julia using ModelingToolkit.jl.

EarthSciMLBase.ICBC

EarthSciMLBase.ICBCcomponent

EarthSciMLBase.constBC

EarthSciMLBase.constIC

EarthSciMLBase.periodicBC

EarthSciMLBase.zerogradBC

Base.:+

Base.:+

EarthSciMLBase.add_dims

EarthSciMLBase.dims

EarthSciMLBase.domains

EarthSciMLBase.ICBC — Type

ICBC(icbc::ICBCcomponent...)

Initial and boundary conditions for a ModelingToolkit.jl PDESystem. It can be used with the + operator to add initial and boundary conditions to a ModelingToolkit.jl ODESystem or Catalyst.jl ReactionSystem.

icbc: The sets of initial and/or boundary conditions.

Example:

using EarthSciMLBase
using ModelingToolkit, DomainSets

# Set up ODE system
@parameters x y t
@variables u(t) v(t)
Dt = Differential(t)

x_min = y_min = t_min = 0.0
x_max = y_max = 1.0
t_max = 11.5

eqs = [
    Dt(u) ~ √abs(v),
    Dt(v) ~ √abs(u),
]

@named sys = ODESystem(eqs)

# Create domains.
indepdomain = t ∈ Interval(t_min, t_max)

partialdomains = [
    x ∈ Interval(x_min, x_max),
    y ∈ Interval(y_min, y_max)
]

# Create constant initial and boundary conditions = 16.0.
icbc = ICBC(constIC(16.0, indepdomain, partialdomains), constBC(16.0, indepdomain, partialdomains))

# Convert to PDESystem and add constant initial and boundary conditions.
pdesys = sys + icbc

pdesys.bcs

# output
10-element Vector{Equation}:
 u(x, y, 0.0) ~ 16.0
 v(x, y, 0.0) ~ 16.0
 u(0.0, y, t) ~ 16.0
 u(1.0, y, t) ~ 16.0
 u(x, 0.0, t) ~ 16.0
 u(x, 1.0, t) ~ 16.0
 v(0.0, y, t) ~ 16.0
 v(1.0, y, t) ~ 16.0
 v(x, 0.0, t) ~ 16.0
 v(x, 1.0, t) ~ 16.0

EarthSciMLBase.ICBCcomponent — Type

Initial and boundary condition components that can be combined to create an ICBC object.

EarthSciMLBase.constBC — Type

Construct constant boundary conditions equal to the value specified by val.

val: The value of the constant boundary conditions.

indepdomain: The independent domain, e.g. t ∈ Interval(t_min, t_max).

partialdomains: The partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

activepartialdomainindex: The indexes of the domains that these boundary conditons are applied to, e.g. 1,2. Leave empty if you want to apply to all partial domains.

EarthSciMLBase.constIC — Type

Construct constant initial conditions equal to the value specified by val.

val: The value of the constant initial conditions.

indepdomain: The independent domain, e.g. t ∈ Interval(t_min, t_max).

partialdomains: The partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

EarthSciMLBase.periodicBC — Type

Construct period boundary conditions for the given partialdomains. Periodic boundary conditions are defined as when the value at one side of the domain is set equal to the value at the other side, so that the domain "wraps around" from one side to the other.

indepdomain: The independent domain, e.g. t ∈ Interval(t_min, t_max).

partialdomains: All of partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

activepartialdomainindex: The indexes of the domains that these boundary conditons are applied to, e.g. 1,2. Leave empty if you want to apply to all partial domains.

EarthSciMLBase.zerogradBC — Type

Construct zero-gradient boundary conditions for the given partialdomains.

indepdomain: The independent domain, e.g. t ∈ Interval(t_min, t_max).

partialdomains: All of partial domains, e.g. [x ∈ Interval(x_min, x_max), y ∈ Interval(y_min, y_max)].

activepartialdomainindex: The indexes of the domains that these boundary conditons are applied to, e.g. 1,2. Leave empty if you want to apply to all partial domains.

Base.:+ — Method

    a::Symbolics.Equation + b::Symbolics.Equation

Add two equations. If the left-hand-sides of the two equations are the same, a single equation is returned with the two right-hand-sides added together. If the left-hand-sides of the two equations are different, a vector of the two equations is returned.

Base.:+ — Method

    a::Vector{Symbolics.Equation} + b::Vector{Symbolics.Equation}

Add two vectors of equations using the same rules as when adding two individual equations.

EarthSciMLBase.add_dims — Method

add_dims(expression, vars, dims)
add_dims(equation, vars, dims)

Add the given dimensions to each variable in vars in the given expression or equation. Each variable in vars must be unidimensional, i.e. defined like @variables u(t) rather than @variables u(..).

Example:

using EarthSciMLBase, ModelingToolkit

@parameters x y k t
@variables u(t) q(t)
exp = 2u + 3k*q + 1
EarthSciMLBase.add_dims(exp, [u, q], [x, y, t])

# output
1 + 2u(x, y, t) + 3k*q(x, y, t)

EarthSciMLBase.dims — Method

dims(icbc)

defined at /home/runner/work/EarthSciMLBase.jl/EarthSciMLBase.jl/src/icbc.jl:250.

Returns the dimensions of the independent and partial domains associated with these initial or boundary conditions.

EarthSciMLBase.domains — Method

domains(icbc)

defined at /home/runner/work/EarthSciMLBase.jl/EarthSciMLBase.jl/src/icbc.jl:258.

Returns the domains associated with these initial or boundary conditions.